Worked on error checking; Worked on AddNewData.Activity; Started on AddNewMajor;
Just about done with error checking just need to change the error text. Added error checking to both AddNewData and Details/Update pages. AddNewData is just about finished just need to tweak GUI and error text. Added a bool to the private static class in the MajorData class so when switching between intents the AddNewData class knows which view to animate for the ViewSwitcher.