Finished work on update functionality; Finished work on save functionality;  Finished on most GUI;  Finished work on formatting data functionality;
Finished working on functions updating the new StudentData to the database. Finished working on save functionality for both NewStudent and NewMajor.  Finished most GUI that was left, just a few tweaks and adjustments remain. Finished making functions to format the data a certain way before saving to the database. Other changes include minor bug fixes related to changing Intents with using ViewSwitcher. I had to use some boolean variables along with a few private static classes to ensure the proper data and views in the ViewSwitcher were showing up. All that's really left is the SeachActivity and the queries needed to perform the proper database searchers.