Finished working on the queries for the Search functions; Finished working on the error checking functions; Finished the Adapter for the Search Results; Created the cell layout for the Search Results list view;
Finished working on the queries for the Search functions; Finished working on the error checking functions; Finished the Adapter for the Search Results; Created the cell layout for the Search Results list view; For the most part I'm just about done with the App. Just need to do some last minute cleaning up of the code and make some comments and comment headers. Run it a few more time to make sure the bugs have been worked out.