Few last minute changes to my arrays in the DatabaseHelper class; Added error checking code to SearchActivity;
Made all my arrays initialize as null at the start of the function calls like what was shown in class. Then added error checking for the null value in my SearchActivity. Had to add an if statement in the SearchResultsAdapter class. The function in this class called getCount() needed to return the length of the search results array but was cause an error since int needed to be returned but if it was null it would cause the error.