Final error tweaking; Fixed error for matching MajorName;
Final error tweaking, added error text view for MainActivity when delete student button is clicked and nothing is selected. Fixed error with MajorName being added even though it already exists. Turns out it was a formatting issue when adding the the names to the database. Just need to update the readme and it's done.